Add files via upload
1. PROBLEM STATEMENT / GOAL
I believe that there is a lot of dependency on loan management which we do solve from loan prediction in the Finance Industry. This is a most sensitive part of finance and we do observe that a lot of people do opt for loans but don’t repay, as a result, we do have to shortlist the ones in whom we should invest in order to minimize the loss.

2. CHARACTERISTICS OF DATA
This dataset has various characteristics which have a lot of diversity and creates it easier to explore it. This has some main columns of gender, employment, dependency, id, marital status, and education. This enables us to see it from a bird's eye view.



Dataset Information

Dream Housing Finance company deals in all home loans. They have a presence across all urban, semi-urban, and rural areas. The customer first applies for a home loan after that company validates the customer's eligibility for a loan. The company wants to automate the loan eligibility process (real-time) based on customer detail provided while filling out the online application form. These details are Gender, Marital Status, Education, Number of Dependents, Income, Loan Amount, Credit History, and others. To automate this process, they have given a problem identifying the customer segments, that are eligible for loan amounts so that they can specifically target these customers.